SurendraDev

Approved-by: Harbans singh

+++ app/Http/Controllers/Api/v1/OrderController.php
@@ -182,7 +182,7 @@ class OrderController extends BaseController
                 $luxury_option = LuxuryOption::where('title', $action)->first();
                 if(checkColumnExists('carts','order_id'))
                 {//get if any order is being edit
-                    $cart = Cart::where('user_id', $user->id)->with(['editingOrder', 'cartvendor'])->first();
+                    $cart = Cart::where('user_id', $user->id)->with(['editingOrder.orderStatusVendor', 'cartvendor'])->first();
                 }else{
                     $cart = Cart::where('user_id', $user->id)->first();
                 }
@@ -223,10 +223,17 @@ class OrderController extends BaseController
                             return $this->errorResponse(__("Order can only be edited before Time limit of ".$order_edit_before_hours." Hours from Scheduled date."), 400);
                         }
                         $VendorOrderStatus = VendorOrderStatus::where('order_id', $order->id)->whereNotIn('order_status_option_id', [1, 2])->count();
-                        if($VendorOrderStatus > 0){
-                            return $this->errorResponse(__("You can not edit this order. Either order is in processed or in processing."), 400);
-                        }
 
+                        $order_vendor_status_error = 0;
+                        foreach ($cart->editingOrder->orderStatusVendor as $key => $status) {
+                            if($status->order_status_option_id  > 2) {
+                                $order_vendor_status_error = 1;
+                            }
+                        }
+                        
+                        if($VendorOrderStatus > 0 || $order_vendor_status_error == 1){
+                            return $this->errorResponse(__("You can not edit this order. Either order is in processed or in processing. Please discard order editing."), 400);
+                        }
                         OrderProduct::where('order_id', $order->id)->delete();
                         OrderProductPrescription::where('order_id', $order->id)->delete();
                         OrderTax::where('order_id', $order->id)->delete();
